WORKAROUND: Qemu Cortex-M4 Thumb entry point
All ARM instructions, Thumb or not, are at least half-word aligned.
Thumb instructions are indicated by the offset having the last bit set
making it appear as though it were quarter-word offset. Qemu
misinterprets this entry point to be the actual Thumb entry point
instead of masking off the last bit of the address as it should. This
workaround places the entry point in non-Thumb space so that it is
half-word aligned and Qemu operates as desired.